Move app to src and implement login / signup
