+++ src/database.py
@@ -1,0 +1,7 @@
+import pymongo
+
+client = pymongo.MongoClient("mongodb://localhost:27017/")
+
+db = client["sauce"]
+
+users = db["users"]
