+++ src/requirements.txt
@@ -1,0 +1,4 @@
+flask
+PyMongo
+pyotp
+bcrypt
